refactor: extract Upgrader to a standalone file for easier testing
